# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CodeFabric Admin API",
        "description": "CodeFabric 管理后台后端接口文档（Laravel 13）",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:4000",
            "description": "本地开发环境"
        }
    ],
    "paths": {
...
        "/api/admin/users": {
            "get": {
                "tags": [
                    "用户管理"
                ],
                "description": "获取用户列表（含角色）",
                "operationId": "adminUserList",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserListEnvelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未登录或登录已失效",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "用户管理"
                ],
                "description": "新增用户",
                "operationId": "adminUserCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "新增成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserItemEnvelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未登录或登录已失效",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "账号已存在",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "参数校验失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/users/{id}": {
            "put": {
                "tags": [
                    "用户管理"
                ],
                "description": "编辑用户",
                "operationId": "adminUserUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "用户 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "编辑成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserItemEnvelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未登录或登录已失效",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "账号已存在",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "参数校验失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "用户管理"
                ],
                "description": "删除用户",
                "operationId": "adminUserDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "用户 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未登录或登录已失效",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "UserRoleBrief": {
                "required": [
                    "id",
                    "name",
                    "code"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "超级管理员"
                    },
                    "code": {
                        "type": "string",
                        "example": "SUPER_ADMIN"
                    }
                },
                "type": "object"
            },
            "UserItem": {
                "required": [
                    "id",
                    "username",
                    "nickname",
                    "status",
                    "createdAt",
                    "roles"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "username": {
                        "type": "string",
                        "example": "admin"
                    },
                    "nickname": {
                        "type": "string",
                        "example": "管理员"
                    },
                    "status": {
                        "type": "string",
                        "example": "ACTIVE",
                        "enum": [
                            "ACTIVE",
                            "DISABLED"
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "example": "2026-04-05 10:00:00"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserRoleBrief"
                        }
                    }
                },
                "type": "object"
            },
            "UserListEnvelope": {
                "required": [
                    "code",
                    "message",
                    "data"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": 0
                    },
                    "message": {
                        "type": "string",
                        "example": "ok"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserItem"
                        }
                    }
                },
                "type": "object"
            },
            "UserItemEnvelope": {
                "required": [
                    "code",
                    "message",
                    "data"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": 0
                    },
                    "message": {
                        "type": "string",
                        "example": "ok"
                    },
                    "data": {
                        "$ref": "#/components/schemas/UserItem"
                    }
                },
                "type": "object"
            },
            "UserCreateRequest": {
                "required": [
                    "username",
                    "nickname",
                    "password",
                    "status"
                ],
                "properties": {
                    "username": {
                        "type": "string",
                        "example": "alice",
                        "maxLength": 50
                    },
                    "nickname": {
                        "type": "string",
                        "example": "Alice",
                        "maxLength": 100
                    },
                    "password": {
                        "type": "string",
                        "example": "123456",
                        "maxLength": 100,
                        "minLength": 6
                    },
                    "status": {
                        "type": "string",
                        "example": "ACTIVE",
                        "enum": [
                            "ACTIVE",
                            "DISABLED"
                        ]
                    },
                    "roleIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": [
                            1,
                            2
                        ]
                    }
                },
                "type": "object"
            },
            "UserUpdateRequest": {
                "properties": {
                    "username": {
                        "type": "string",
                        "example": "alice",
                        "maxLength": 50
                    },
                    "nickname": {
                        "type": "string",
                        "example": "Alice",
                        "maxLength": 100
                    },
                    "password": {
                        "type": "string",
                        "example": "new_123456",
                        "nullable": true,
                        "maxLength": 100,
                        "minLength": 6
                    },
                    "status": {
                        "type": "string",
                        "example": "DISABLED",
                        "enum": [
                            "ACTIVE",
                            "DISABLED"
                        ]
                    },
                    "roleIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": [
                            2
                        ]
                    }
                },
                "type": "object"
            },
...
            "ApiEnvelope": {
                "required": [
                    "code",
                    "message",
                    "data"
                ],
                "properties": {
                    "code": {
                        "description": "CodeFabric 后端 OpenAPI 根定义。\n\n说明：\n- 使用 swagger-php Attributes 维护接口文档。\n- 业务返回统一为 { code, message, data }，通过组件 Schema 复用。",
                        "type": "integer",
                        "example": 0
                    },
                    "message": {
                        "type": "string",
                        "example": "ok"
                    },
                    "data": {
                        "nullable": true,
                        "oneOf": [
                            {
                                "type": "object",
                                "additionalProperties": true
                            },
                            {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "additionalProperties": true
                                }
                            },
                            {
                                "type": "string"
                            },
                            {
                                "type": "integer"
                            },
                            {
                                "type": "boolean"
                            }
                        ]
                    }
                },
                "type": "object"
            },
            "ApiErrorEnvelope": {
                "required": [
                    "code",
                    "message",
                    "data"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": 2004
                    },
                    "message": {
                        "type": "string",
                        "example": "请求失败"
                    },
                    "data": {
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "在 Authorization 请求头中传入 Bearer Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "cookieAuth": {
                "type": "apiKey",
                "description": "浏览器 Cookie 鉴权",
                "name": "admin_token",
                "in": "cookie"
            }
        }
    },
...
}